furiosa_mapping

Trait AtomExt

pub trait AtomExt {
    // Required methods
    fn size(&self) -> usize;
    fn idents(&self) -> RVec<Ident>;
    fn contains_ident(&self, ident: Ident) -> bool;
    fn find_symbol_size(&self, ident: Ident) -> ROption<usize>;
}

Methods for Atom.

Required Methods

fn size(&self) -> usize

Returns the size of the atomic mapping expression.

fn idents(&self) -> RVec<Ident>

Returns the idents contained in this atom.

fn contains_ident(&self, ident: Ident) -> bool

Returns true if this atom contains the given ident (recursively).

fn find_symbol_size(&self, ident: Ident) -> ROption<usize>

Returns the original declared size of the given ident’s Symbol.