furiosa_mapping

Trait IndexExt

pub trait IndexExt: Sized {
    // Required methods
    fn new() -> Self;
    fn add(&mut self, other: Index);
    fn mark_invalid(&mut self);
    fn add_term(&mut self, term: Term, value: usize);
    fn add_mapping<I: M>(&mut self, value: usize);
    fn ident_value(&self, ident: Ident) -> RResult<usize, IndexValueError>;
    fn finalize(self) -> RResult<RSortedMap<Term, usize>, ()>;
    fn gen_indexes(&self, mapping: FMapping) -> RVec<Index>;
}

Methods for Index.

Required Methods

fn new() -> Self

Creates a new empty index.

fn add(&mut self, other: Index)

Adds another index to this index.

fn mark_invalid(&mut self)

Marks this index as invalid.

fn add_term(&mut self, term: Term, value: usize)

Adds a term to this index.

fn add_mapping<I: M>(&mut self, value: usize)

Adds a mapping to this index.

fn ident_value(&self, ident: Ident) -> RResult<usize, IndexValueError>

Returns the contribution of a specific ident to this Index.

fn finalize(self) -> RResult<RSortedMap<Term, usize>, ()>

Finalizes the index.

fn gen_indexes(&self, mapping: FMapping) -> RVec<Index>

Generates all possible indexes based on the given mapping.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.